pyproject.toml
[project]
name = "Musicology"
version = "0.0.1"
description = "Langgraph based app to recommend music based on user input"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "google-generativeai",
    "ipykernel",
    "langchain-community",
    "langchain-google-genai",
    "langchain[anthropic,openai]",
    "langgraph",
    "pandas",
    "python-dotenv",
    "langgraph-cli[inmem]",
    "spotipy>=2.25.2",
    "wikipedia>=1.4.0",
]

